# packages/contracts/generated/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BeOnEdge API",
    "version": "v1"
  },
...
  "paths": {
...
    "/v1/activations/complete": {
      "post": {
        "operationId": "completeActivation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9_-]{43}$/u"
                  },
                  "password": {
                    "type": "string"
                  },
                  "device": {
                    "type": "object",
                    "properties": {
                      "installationId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "name": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 80
                      },
                      "platform": {
                        "type": "string",
                        "enum": [
                          "android"
                        ]
                      },
                      "appVersion": {
                        "type": "string",
                        "pattern": "^[0-9]+[.][0-9]+[.][0-9]+([+-][A-Za-z0-9.-]+)?$/u"
                      }
                    },
                    "required": [
                      "installationId",
                      "name",
                      "platform",
                      "appVersion"
                    ],
                    "additionalProperties": false
                  }
                },
                "required": [
                  "token",
                  "password",
                  "device"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "userId": {
                              "type": "string",
                              "format": "uuid"
                            },
                            "fullName": {
                              "type": "string"
                            },
                            "email": {
                              "type": "string",
                              "maxLength": 254,
                              "format": "email"
                            },
                            "phoneMasked": {
                              "type": "string",
                              "pattern": "^\\+[1-9][0-9]{0,2}[*]{6}[0-9]{4}$/u"
                            },
                            "accountStatus": {
                              "type": "string",
                              "enum": [
                                "active"
                              ]
                            }
                          },
                          "required": [
                            "userId",
                            "fullName",
                            "email",
                            "phoneMasked",
                            "accountStatus"
                          ],
                          "additionalProperties": false
                        },
                        "accessToken": {
                          "type": "string",
                          "minLength": 100,
                          "maxLength": 4096
                        },
                        "accessTokenExpiresAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "refreshToken": {
                          "type": "string",
                          "pattern": "^[A-Za-z0-9_-]{43}$/u"
                        },
                        "refreshTokenExpiresAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "sessionId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      },
                      "required": [
                        "user",
                        "accessToken",
                        "accessTokenExpiresAt",
                        "refreshToken",
                        "refreshTokenExpiresAt",
                        "sessionId"
                      ],
                      "additionalProperties": false
                    },
                    "error": {
                      "type": "null"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "idempotencyReplay": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "requestId",
                        "timestamp"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "ok",
                    "data",
                    "error",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "410": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "415": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "x-app-version",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-client-platform",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/auth/native/login": {
      "post": {
        "operationId": "nativeLogin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "maxLength": 254,
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  },
                  "device": {
                    "type": "object",
                    "properties": {
                      "installationId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "name": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 80
                      },
                      "platform": {
                        "type": "string",
                        "enum": [
                          "android"
                        ]
                      },
                      "appVersion": {
                        "type": "string",
                        "pattern": "^[0-9]+[.][0-9]+[.][0-9]+([+-][A-Za-z0-9.-]+)?$/u"
                      }
                    },
                    "required": [
                      "installationId",
                      "name",
                      "platform",
                      "appVersion"
                    ],
                    "additionalProperties": false
                  }
                },
                "required": [
                  "email",
                  "password",
                  "device"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "userId": {
                              "type": "string",
                              "format": "uuid"
                            },
                            "fullName": {
                              "type": "string"
                            },
                            "email": {
                              "type": "string",
                              "maxLength": 254,
                              "format": "email"
                            },
                            "phoneMasked": {
                              "type": "string",
                              "pattern": "^\\+[1-9][0-9]{0,2}[*]{6}[0-9]{4}$/u"
                            },
                            "accountStatus": {
                              "type": "string",
                              "enum": [
                                "active"
                              ]
                            }
                          },
                          "required": [
                            "userId",
                            "fullName",
                            "email",
                            "phoneMasked",
                            "accountStatus"
                          ],
                          "additionalProperties": false
                        },
                        "accessToken": {
                          "type": "string",
                          "minLength": 100,
                          "maxLength": 4096
                        },
                        "accessTokenExpiresAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "refreshToken": {
                          "type": "string",
                          "pattern": "^[A-Za-z0-9_-]{43}$/u"
                        },
                        "refreshTokenExpiresAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "sessionId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      },
                      "required": [
                        "user",
                        "accessToken",
                        "accessTokenExpiresAt",
                        "refreshToken",
                        "refreshTokenExpiresAt",
                        "sessionId"
                      ],
                      "additionalProperties": false
                    },
                    "error": {
                      "type": "null"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "idempotencyReplay": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "requestId",
                        "timestamp"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "ok",
                    "data",
                    "error",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "415": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "x-app-version",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-client-platform",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/auth/native/refresh": {
      "post": {
        "operationId": "refreshNativeSession",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9_-]{43}$/u"
                  },
                  "rotationId": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": [
                  "refreshToken",
                  "rotationId"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string",
                          "minLength": 100,
                          "maxLength": 4096
                        },
                        "accessTokenExpiresAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "refreshToken": {
                          "type": "string",
                          "pattern": "^[A-Za-z0-9_-]{43}$/u"
                        },
                        "refreshTokenExpiresAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "sessionId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      },
                      "required": [
                        "accessToken",
                        "accessTokenExpiresAt",
                        "refreshToken",
                        "refreshTokenExpiresAt",
                        "sessionId"
                      ],
                      "additionalProperties": false
                    },
                    "error": {
                      "type": "null"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "idempotencyReplay": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "requestId",
                        "timestamp"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "ok",
                    "data",
                    "error",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "415": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "x-app-version",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-client-platform",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/v1/auth/native/logout": {
      "post": {
        "operationId": "logoutNativeSession",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "pattern": "^[A-Za-z0-9_-]{43}$/u"
                  }
                },
                "required": [
                  "refreshToken"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "loggedOut": {
                          "type": "boolean",
                          "enum": [
                            true
                          ]
                        }
                      },
                      "required": [
                        "loggedOut"
                      ],
                      "additionalProperties": false
                    },
                    "error": {
                      "type": "null"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "idempotencyReplay": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "requestId",
                        "timestamp"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "ok",
                    "data",
                    "error",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "415": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "x-app-version",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-client-platform",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    }
  },
...
}